Conflicts:
	RandomFruit/app/views/dash/modals/createaticket.blade.php

+++ RandomFruit/app/database/migrations/2014_03_26_140430_create_tickets_table.php
@@ -19,8 +19,8 @@ class CreateTicketsTable extends Migration {
 			$table->timestamps();
 			$table->string('title');
 			$table->mediumText('description');
-			$table->integer('planned_hours');
-			$table->integer('actual_hours');
+			$table->decimal('planned_hours', 6, 1);
+			$table->decimal('actual_hours', 6, 1);
 			$table->integer('number');
 			$table->date('due_date')->nullable();
 			$table->integer('project_id')->unsigned();
